Separated InlineTaskForm component for inline edit
